#62 ES6 Refactor Still requires additional clean up
