+++ packages/livedoc-mocha/.vscode/settings.json
@@ -4,5 +4,8 @@
     "typescript.tsdk": "./node_modules/typescript/lib",
     "markdown.styles": [
         ".vscode/markdown.css"
+    ],
+    "cSpell.words": [
+        "Polyfils"
     ]
 }
