+++ packages/livedoc-mocha/_src/app/BackgroundContext.ts
@@ -1,0 +1,4 @@
+import { ScenarioContext } from "./ScenarioContext";
+
+export class BackgroundContext extends ScenarioContext {
+}
