+++ packages/livedoc-mocha/_src/app/CommandLineOptions.ts
@@ -1,0 +1,5 @@
+export class CommandLineOptions {
+    public include: string[] = [];
+    public exclude: string[] = [];
+    public showFilterConflicts: boolean = false;
+}
