+++ packages/livedoc-mocha/_src/app/ScenarioContext.ts
@@ -1,0 +1,12 @@
+import { StepContext } from "./StepContext";
+
+export class ScenarioContext {
+    title: string;
+    description: string;
+    given: StepContext;
+    //givens: StepContext[] = [];
+    //when: StepContext;
+    //whens: StepContext[] = [];
+    and: StepContext[] = [];
+    tags: string[];
+}
