+++ packages/livedoc-mocha/_src/app/ScenarioOutlineContext.ts
@@ -1,0 +1,6 @@
+import { ScenarioContext } from "./ScenarioContext";
+
+export class ScenarioOutlineContext extends ScenarioContext {
+    public example: DataTableRow;
+    public exampleRaw: DataTableRow;
+}
