+++ packages/livedoc-mocha/_src/app/model/Describe.ts
@@ -1,0 +1,10 @@
+import { Test } from "./Test";
+
+// Legacy BDD model
+export class Describe {
+    constructor (public title: string) {
+
+    }
+    public children: Describe[] = [];
+    public tests: Test[] = [];
+}
