+++ packages/livedoc-mocha/_src/app/model/Feature.ts
@@ -1,0 +1,40 @@
+import { LiveDocDescribe } from "./LiveDocDescribe";
+import { Background } from "./Background";
+import { Scenario } from "./Scenario";
+import { FeatureContext } from "../FeatureContext";
+import { BackgroundContext } from "../BackgroundContext";
+
+export class Feature extends LiveDocDescribe {
+    public filename: string;
+    public background: Background;
+    public scenarios: Scenario[] = [];
+
+    public executionTime: number;
+
+    constructor () {
+        super()
+        this.displayPrefix = "Feature";
+        this.displayIndentLength = 4;
+    }
+
+
+    public getFeatureContext(): FeatureContext {
+        return ({
+            filename: this.filename,
+            title: this.title,
+            description: this.description,
+            tags: this.tags
+        });
+    }
+
+    public getBackgroundContext(): BackgroundContext {
+        const context = this.background.getScenarioContext();
+        return ({
+            title: context.title,
+            description: context.description,
+            given: context.given,
+            and: context.and,
+            tags: context.tags
+        });
+    }
+}
