+++ packages/livedoc-mocha/_src/app/model/LiveDocRuleOption.ts
@@ -1,0 +1,5 @@
+export enum LiveDocRuleOption {
+    enabled,
+    disabled,
+    warning
+}
