+++ packages/livedoc-mocha/_src/app/model/Table.ts
@@ -1,0 +1,5 @@
+export class Table {
+    public name: string = "";
+    public description: string = "";
+    public dataTable: DataTableRow[] = [];
+}
