+++ packages/livedoc-mocha/_src/app/model/Test.ts
@@ -1,0 +1,5 @@
+export class Test {
+    constructor (public title: string) {
+
+    }
+}
