+++ packages/livedoc-mocha/cSpell.json
@@ -6,7 +6,8 @@
     "language": "en",
     // words - list of words to be always considered correct
     "words": [
-        "livedoc"
+        "livedoc",
+        "Polyfils"
     ],
     // flagWords - list of words to be always considered incorrect
     // This is useful for offensive words and common spelling errors.
