+++ packages/livedoc-mocha/tsconfig.json
@@ -8,7 +8,13 @@
         ],
         "outDir": "build",
         "declaration": true,
-        "removeComments": true
+        "removeComments": true,
+        "sourceMap": true,
+        "alwaysStrict": true,
+        "forceConsistentCasingInFileNames": true,
+        "noImplicitThis": true,
+        "noUnusedLocals": true,
+        "pretty": true
     },
     "include": [
         "_src/app/**/*.ts",
